requirements.txt
asgiref==3.4.1
black==21.12b0
certifi==2021.10.8
click==8.0.3
coverage==6.2
Django==3.2.5
djangorestframework==3.13.1
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.8.6
pytz==2021.3
sqlparse==0.4.1
tomli==1.2.3
typing-extensions==3.10.0.2
